move conversions to topicDto
